backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
alembic==1.13.1
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-multipart==0.0.9
passlib==1.7.4
python-jose==3.3.0
chromadb==0.4.24
langchain-text-splitters==0.2.0
google-generativeai==0.7.0
uvicorn==0.29.0
SQLAlchemy==2.0.28
asyncpg==0.29.0
pypdf==4.2.0
